Abstract the constructor to parent class
